Implemented showing restaurant details logic

+++ app/src/main/java/com/demo/nerdeyesem/data/entities/responses/UserRatingResponse.kt
@@ -3,7 +3,7 @@ package com.demo.nerdeyesem.data.entities.responses
 import com.google.gson.annotations.SerializedName
 
 data class UserRatingResponse(
-    @field:SerializedName("aggregate_rating") val aggregateRating: Double,
+    @field:SerializedName("aggregate_rating") val aggregateRating: Float,
     @field:SerializedName("rating_text") val ratingText: String,
     @field:SerializedName("rating_color") val ratingColor: String,
     @field:SerializedName("rating_obj") val ratingObj: RatingObjResponse,
